+++ docs/dev-docs/changelog.rst
@@ -5,6 +5,7 @@ Upcoming
 --------
 
 - Added Rust crate for wrapping kubos-hal-iobc.
+- Added iOBC supervisor graphql service.
 
 v1.1.0 - Jan 24 2018
 --------------------